+++ complete.html
@@ -2,9 +2,20 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Title</title>
+    <link rel="stylesheet" href="./style/style.css">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Lato&family=Tinos:wght@400;700&display=swap" rel="stylesheet">
+    <title>Complete</title>
 </head>
-<body>
-
+<body class="checkout_form">
+<div class="result_wrapper">
+    <div class="payment_wrapper">
+        <div class="payment_inner" id="submit">
+        </div>
+    </div>
+</div>
+<script type="module" src="complete.js"></script>
 </body>
 </html>
