--- delivery_form.html
@@ -1,91 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <link rel="stylesheet" href="style.css">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Lato&family=Tinos:wght@400;700&display=swap" rel="stylesheet">
-    <title>Checkout</title>
-</head>
-<body class="checkout" >
-    <div class="wrapper">
-        <div class="payment_wrapper">
-            <div class="payment_inner">
-                <div class="payment_logo">
-                    <span class="logo"></span>
-                    <h2 class="book_title">Checkout form</h2>
-                </div>
-                <div class="cart_container"></div>
-                <form action="/complete.html" method="get" class="delivery" id="delivery">
-                    <div class="text_form">
-                        <div class="input_element">
-                            <label for="input-name">Name</label>
-                            <input id="input-name" pattern="[A-Za-z]{4,}" required class="validate" autocomplete="off">
-                        </div>
-                        <div class="input_element">
-                            <label for="input-surname">Surname</label>
-                            <input id="input-surname" pattern="[A-Za-z]{5,}" required class="validate" autocomplete="off">
-                        </div>
-                        <div class="input_element">
-                            <label for="input-date">Date</label>
-                            <input type="date" id="input-date" min='' required class="validate" autocomplete="off">
-                        </div>
-                        <div class="input_element">
-                            <label for="input-street">Street</label>
-                            <input id="input-street" pattern="[A-Za-z]{5,}" required class="validate" autocomplete="off">
-                        </div>
-                        <div class="input_element">
-                            <label for="input-house">House №</label>
-                            <input id="input-house" minlength="1" required class="validate" autocomplete="off">
-                        </div>
-                        <div class="input_element">
-                            <label for="input-flat">Flat №</label>
-                            <input id="input-flat" minlength="1" required class="validate" autocomplete="off">
-                        </div>
-                    </div>
-
-                    <div class="options">
-                        <div class="gift" id="gift" >
-                            <h4 class="subtitle">2 gift: (optional)</h4>
-                            <div class="inner_element"><label for="pack">pack as a gift</label>
-                                <input class="checkbox" type="checkbox" id="pack" name="gift" value="" required >
-                            </div>
-                            <div class="inner_element"><label for="postcard">add postcard</label>
-                                <input class="checkbox" type="checkbox" id="postcard" name="gift" value="" required>
-                            </div>
-                            <div class="inner_element"><label for="discount">2% discount to the next time</label>
-                                <input class="checkbox" type="checkbox" id="discount" name="gift" value="" required>
-                            </div>
-                            <div class="inner_element"><label for="pencil">branded pencil</label>
-                                <input class="checkbox" type="checkbox" id="pencil" name="gift" value="" required>
-                            </div>
-                        </div>
-
-                        <div class="payment_option">
-                            <h4 class="subtitle">Choose payment option:</h4>
-                            <div class="input_card">
-                                <label for="Cash">Cash</label>
-                                <input type="radio" id="Cash" name="payment" value="Cash" required>
-                            </div>
-                            <div class="input_card">
-                                <label for="Card">Card</label>
-                                <input type="radio" id="Card" name="payment" value="Card" required>
-                            </div>
-                        </div>
-                    </div>
-
-                    <div class="buttons">
-                        <button class="button_cancel" onclick="location.href='index.html'"><p class="button_text">Cancel</p></button>
-                        <button class="button_submit" id="button_submit" disabled><p class="button_text" >Submit</p></button>
-                    </div>
-                </form>
-            </div>
-        </div>
-
-    </div>
-    <script type="text/javascript" src="variables.js"></script>
-    <script type="text/javascript" src="checkout.js"></script>
-</body>
-</html>
