+++ index.html
@@ -2,21 +2,19 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="./style/style.css">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Lato&family=Tinos:wght@400;700&display=swap" rel="stylesheet">
     <title>Book_shop</title>
 </head>
-<body class="body">
-<div class="wrapper"></div>
+    <body class="body">
+    <div class="wrapper"></div>
+    <script type="module" src="script/script.js"></script>
+    <script type="module" src="script/modal.js"></script>
+    <script type="module" src="script/drag.js"></script>
+    <script type="module" src="script/checkout.js"></script>
 
-<script type="text/javascript" src="script.js"></script>
-<script type="text/javascript" src="modal.js"></script>
-<script type="text/javascript" src="variables.js"></script>
-<script type="text/javascript" src="drag.js"></script>
-
-
-</body>
+    </body>
 </html>
