+++ style/checkout.css
@@ -1,11 +1,16 @@
+.checkout_form{
+    background-color: #FFD25A;
+    background-size: cover;
+}
+
 .total_wrapper{
     display: flex;
     flex-direction: row;
     justify-content: space-between;
     align-items: center;
     margin: 0 20px 10px 20px;
-    background-color: #bed7b8;
-    border-radius: 30px;
+    background-color: #FFD25A;
+    border-radius: 9px;
 }
 
 .total_count{
@@ -35,32 +40,35 @@
     display: flex;
     flex-direction: row;
     align-items: center;
-    background-color: #C5E5A5;
     border-radius: 100px;
     padding: 4px;
 }
 
 .checkout{
     background-color: #F9D876;
+    margin-top: 30%;
 }
 
 .payment_wrapper{
-    width: 600px;
-    height: 600px;
+    width: 630px;
     background-color: #F6F6F6;
     border-radius: 9px;
+    margin-top: 20%;
 }
 
 .payment_inner{
     display: flex;
     flex-direction: column;
     justify-content: center;
+    margin: 10px;
+    gap: 10px;
 }
 
 .payment_logo{
     display: flex;
     flex-direction: row;
     align-items: center;
+    margin-left: 20px;
 }
 .delivery{
     display: flex;
@@ -82,6 +90,12 @@
 .cart_container{
     display: flex;
     flex-direction: column;
+    gap: 10px;
+    background: white;
+    border-radius: 9px;
+    padding: 20px;
+    max-height: 550px;
+    overflow: scroll;
 }
 
 .input_element{
@@ -96,9 +110,9 @@
     font-size: 17px;
     line-height: 22.1px;
     letter-spacing: 0.06em;
-    border-radius: 100px;
+    border-radius: 9px;
     cursor: pointer;
-    background: #C5E5A5;
+    background: #FFD25A;
     color: #292929;
     padding: 5px 10px 5px 10px;
     border: none;
@@ -111,9 +125,9 @@
     font-size: 17px;
     line-height: 22.1px;
     letter-spacing: 0.06em;
-    border-radius: 100px;
+    border-radius: 9px;
     cursor: pointer;
-    background: #C5E5A5;
+    background: #FFD25A;
     color: #292929;
     padding: 5px 10px 5px 10px;
     border: none;
@@ -124,29 +138,30 @@
 }
 
 button.button_submit:hover{
-    background: #98B66E;
+    background: #FFAA5A;
     transition: all .3s ease-out;
 }
 
 button.button_submit:disabled{
-    background: #98B66E;
+    background: #FF785A;
+    opacity: 0.5;
+    cursor: default;
 }
 
 button.button_cancel:hover{
-    background: #98B66E;
+    background: #FFAA5A;
     transition: all .3s ease-out;
 }
 
 input:invalid:required {
-    background-color: #98B66E;
+    border: 1px solid #FF785A;
 }
-.inner_delivery{
 
-}
+
 .text_form{
     display: flex;
     flex-wrap: wrap;
-    gap: 5px;
+    gap: 10px;
     justify-content: center;
 }
 
@@ -160,7 +175,6 @@ input:invalid:required {
     width: 150px;
     height: 10px;
     padding: 10px 10px 10px 10px;
-    background-color: #C5E5A5;
     border-radius: 9px;
     border: none;
 }
@@ -183,9 +197,7 @@ label{
     margin: 15px 0 5px 0;
 }
 
-.payment_option{
 
-}
 
 .options{
     display: flex;
@@ -195,5 +207,40 @@ label{
     display: flex;
     gap: 5px;
     justify-content: end;
-    margin-right: 50px;
+    margin-right: 20px;
+}
+
+.item{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
+.item_name{
+    margin: 0;
+    font-family: 'Georgia', sans-serif;
+    font-style: normal;
+    font-weight: 400;
+    font-size: 14px;
+    line-height: 22.1px;
+    letter-spacing: 0.06em;
+}
+
+.item_price{
+    margin: 0;
+    font-family: 'Georgia', sans-serif;
+    font-style: normal;
+    font-weight: 400;
+    font-size: 14px;
+    line-height: 22.1px;
+    letter-spacing: 0.06em;
+}
+
+.cart_total{
+    display: flex;
+    justify-content: end;
+    margin-right: 20px;
+}
+
+input:invalid{
+
 }
