+++ style/header.css
@@ -1,5 +1,5 @@
 .header{
-    background-color: #F9D876;
+    background-color: #FFD25A;;
     background-size: cover;
     gap: 154px;
     height: 1023px;
@@ -18,6 +18,7 @@
     display: flex;
     flex-direction: column;
     margin-left: 20px;
+    gap: 3px;
 }
 
 .header_title{
