+++ style/modal.css
@@ -100,6 +100,6 @@
     margin-top: 2px;
     width: 12px;
     height: 12px;
-    background-image: url("icons/modal_close_button_Vector.svg");
+    background-image: url("../icons/modal_close_button_Vector.svg");
     display: inline-flex;
 }
