--- variables.js
@@ -1,10 +1,0 @@
-let start = '';
-let access = '';
-
-let data = '';
-let result = 0;
-
-
-let info = '';
-
-
